requirements.txt
Django==5.0.3
django-allauth==0.61.1
django-debug-toolbar==4.3.0
django-import-export==3.3.7
flake8==7.0.0
flake8-print==5.0.0
Markdown==3.6
mypy==1.9.0
mypy-extensions==1.0.0
types-Markdown==3.6.0.20240316
